re-organize RecoLocalTracker/SiStripClusterizer/test/ plugins
