+++ RecoLocalTracker/SiStripClusterizer/test/ClusterRefinerTagMCmerged.cc
@@ -1,6 +1,42 @@
-#include "RecoLocalTracker/SiStripClusterizer/test/ClusterRefinerTagMCmerged.h"
+//
+// Use MC truth to identify merged clusters, i.e., those associated with more than one
+// (in-time) SimTrack.
+//
+// Author:  Bill Ford (wtford)  6 March 2015
+//
+
+// system includes
+#include <memory>
+
+// user includes
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+#include "SimTracker/TrackerHitAssociation/interface/TrackerHitAssociator.h"
 #include "DataFormats/SiStripCluster/interface/SiStripCluster.h"
 #include "FWCore/Framework/interface/Event.h"
+
+class ClusterRefinerTagMCmerged : public edm::stream::EDProducer<> {
+public:
+  explicit ClusterRefinerTagMCmerged(const edm::ParameterSet& conf);
+  virtual void produce(edm::Event&, const edm::EventSetup&);
+
+private:
+  template <class T>
+  bool findInput(const edm::EDGetTokenT<T>&, edm::Handle<T>&, const edm::Event&);
+  virtual void refineCluster(const edm::Handle<edmNew::DetSetVector<SiStripCluster>>& input,
+                             edmNew::DetSetVector<SiStripCluster>& output);
+
+  const edm::InputTag inputTag;
+  typedef edm::EDGetTokenT<edmNew::DetSetVector<SiStripCluster>> token_t;
+  token_t inputToken;
+  edm::ParameterSet confClusterRefiner_;
+  bool useAssociateHit_;
+
+  TrackerHitAssociator::Config trackerHitAssociatorConfig_;
+  std::unique_ptr<TrackerHitAssociator> associator_;
+};
 
 ClusterRefinerTagMCmerged::ClusterRefinerTagMCmerged(const edm::ParameterSet& conf)
     : inputTag(conf.getParameter<edm::InputTag>("UntaggedClusterProducer")),
@@ -87,3 +123,6 @@ inline bool ClusterRefinerTagMCmerged::findInput(const edm::EDGetTokenT<T>& tag,
   e.getByToken(tag, handle);
   return handle.isValid();
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(ClusterRefinerTagMCmerged);
