+++ RecoLocalTracker/SiStripClusterizer/test/ClusterizerUnitTesterESProducer.cc
@@ -1,5 +1,45 @@
-#include "RecoLocalTracker/SiStripClusterizer/test/ClusterizerUnitTesterESProducer.h"
+// system includes
+#include <memory>
+
+// user includes
+#include "CalibFormats/SiStripObjects/interface/SiStripGain.h"
+#include "CalibFormats/SiStripObjects/interface/SiStripQuality.h"
+#include "CalibTracker/Records/interface/SiStripGainRcd.h"
+#include "CalibTracker/Records/interface/SiStripQualityRcd.h"
 #include "CalibTracker/SiStripCommon/interface/SiStripDetInfoFileReader.h"
+#include "CondFormats/DataRecord/interface/SiStripNoisesRcd.h"
+#include "CondFormats/SiStripObjects/interface/SiStripApvGain.h"
+#include "CondFormats/SiStripObjects/interface/SiStripNoises.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+
+class ClusterizerUnitTesterESProducer : public edm::ESProducer {
+  typedef edm::ParameterSet PSet;
+  typedef std::vector<PSet> VPSet;
+  typedef VPSet::const_iterator iter_t;
+
+public:
+  ClusterizerUnitTesterESProducer(const PSet&);
+  ~ClusterizerUnitTesterESProducer() {}
+  std::shared_ptr<const SiStripGain> produceGainRcd(const SiStripGainRcd&) { return gain_; }
+  std::shared_ptr<const SiStripNoises> produceNoisesRcd(const SiStripNoisesRcd&) { return noises_; }
+  std::shared_ptr<const SiStripQuality> produceQualityRcd(const SiStripQualityRcd&) { return quality_; }
+
+private:
+  void extractNoiseGainQuality(const PSet&, SiStripQuality*, SiStripApvGain*, SiStripNoises*);
+  void extractNoiseGainQualityForDetId(uint32_t, const VPSet&, SiStripQuality*, SiStripApvGain*, SiStripNoises*);
+
+  void setNoises(uint32_t, std::vector<std::pair<uint16_t, float> >&, SiStripNoises*);
+  void setGains(uint32_t, std::vector<std::pair<uint16_t, float> >&, SiStripApvGain*);
+
+  // These objects might be shared across multiple concurrent
+  // IOVs and are not allowed to be modified after the module
+  // constructor finishes.
+  std::unique_ptr<const SiStripApvGain> apvGain_;
+  std::shared_ptr<const SiStripGain> gain_;
+  std::shared_ptr<const SiStripNoises> noises_;
+  std::shared_ptr<const SiStripQuality> quality_;
+};
 
 ClusterizerUnitTesterESProducer::ClusterizerUnitTesterESProducer(const edm::ParameterSet& conf) {
   const auto detInfo =
@@ -99,3 +139,10 @@ void ClusterizerUnitTesterESProducer::setGains(uint32_t detId,
     throw cms::Exception("Trying to set gain twice for same detId: ") << detId;
   return;
 }
+
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/Utilities/interface/typelookup.h"
+#include "FWCore/Framework/interface/eventsetuprecord_registration_macro.h"
+DEFINE_FWK_EVENTSETUP_MODULE(ClusterizerUnitTesterESProducer);
