+++ RecoLocalTracker/SiStripClusterizer/test/SiStripApproximatedClustersDump.cc
@@ -1,4 +1,58 @@
-#include "RecoLocalTracker/SiStripClusterizer/test/SiStripApproximatedClustersDump.h"
+// system includes
+#include <memory>
+#include <iostream>
+
+// user include files
+#include "CommonTools/UtilAlgos/interface/TFileService.h"
+#include "CommonTools/Utils/interface/TFileDirectory.h"
+#include "DataFormats/Common/interface/DetSet.h"
+#include "DataFormats/Common/interface/DetSetVector.h"
+#include "DataFormats/Common/interface/DetSetVectorNew.h"
+#include "DataFormats/SiStripCluster/interface/SiStripApproximateCluster.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
+#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/ParameterSet/interface/ParameterSetDescription.h"
+#include "FWCore/ServiceRegistry/interface/Service.h"
+
+//ROOT inclusion
+#include "TROOT.h"
+#include "TFile.h"
+#include "TNtuple.h"
+#include "TTree.h"
+#include "TMath.h"
+#include "TList.h"
+#include "TString.h"
+
+//
+// class decleration
+//
+
+class SiStripApproximatedClustersDump : public edm::one::EDAnalyzer<edm::one::SharedResources> {
+public:
+  explicit SiStripApproximatedClustersDump(const edm::ParameterSet&);
+  ~SiStripApproximatedClustersDump() override;
+
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  void analyze(const edm::Event&, const edm::EventSetup&) override;
+
+  edm::InputTag inputTagClusters;
+  edm::EDGetTokenT<edmNew::DetSetVector<SiStripApproximateCluster>> clusterToken;
+
+  TTree* outNtuple;
+  edm::Service<TFileService> fs;
+
+  uint32_t detId;
+  uint16_t barycenter;
+  uint16_t width;
+  uint8_t avCharge;
+  edm::EventNumber_t eventN;
+};
 
 SiStripApproximatedClustersDump::SiStripApproximatedClustersDump(const edm::ParameterSet& conf) {
   inputTagClusters = conf.getParameter<edm::InputTag>("approximatedClustersTag");
@@ -14,7 +68,7 @@ SiStripApproximatedClustersDump::SiStripApproximatedClustersDump(const edm::Para
   outNtuple->Branch("charge", &avCharge, "charge/b");
 }
 
-SiStripApproximatedClustersDump::~SiStripApproximatedClustersDump() {}
+SiStripApproximatedClustersDump::~SiStripApproximatedClustersDump() = default;
 
 void SiStripApproximatedClustersDump::analyze(const edm::Event& event, const edm::EventSetup& es) {
   edm::Handle<edmNew::DetSetVector<SiStripApproximateCluster>> clusterCollection = event.getHandle(clusterToken);
@@ -37,3 +91,6 @@ void SiStripApproximatedClustersDump::fillDescriptions(edm::ConfigurationDescrip
   desc.add<edm::InputTag>("approximatedClustersTag", edm::InputTag("SiStripClusters2ApproxClusters"));
   descriptions.add("SiStripApproximatedClustersDump", desc);
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(SiStripApproximatedClustersDump);
