+++ RecoLocalTracker/SiStripClusterizer/test/StripByStripTestDriver.cc
@@ -1,22 +1,42 @@
+// system includes
+#include <memory>
 
-#include "RecoLocalTracker/SiStripClusterizer/test/StripByStripTestDriver.h"
-#include "DataFormats/SiStripDigi/interface/SiStripDigi.h"
+// user includes
 #include "DataFormats/SiStripCluster/interface/SiStripCluster.h"
+#include "DataFormats/SiStripDigi/interface/SiStripDigi.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
 #include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+#include "RecoLocalTracker/SiStripClusterizer/interface/StripClusterizerAlgorithm.h"
+#include "RecoLocalTracker/SiStripClusterizer/interface/StripClusterizerAlgorithmFactory.h"
+
+class StripByStripTestDriver : public edm::stream::EDProducer<> {
+  typedef edmNew::DetSetVector<SiStripCluster> output_t;
+
+public:
+  StripByStripTestDriver(const edm::ParameterSet&);
+  ~StripByStripTestDriver();
+
+private:
+  void produce(edm::Event&, const edm::EventSetup&);
+
+  const edm::InputTag inputTag;
+  const bool hlt;
+
+  std::unique_ptr<StripClusterizerAlgorithm> algorithm;
+};
 
 StripByStripTestDriver::StripByStripTestDriver(const edm::ParameterSet& conf)
-    : inputTag(conf.getParameter<edm::InputTag>("DigiProducer")),
-      hlt(conf.getParameter<bool>("HLT"))  //,
-/*hltFactory(0)*/ {
+    : inputTag(conf.getParameter<edm::InputTag>("DigiProducer")), hlt(conf.getParameter<bool>("HLT")) {
   algorithm = StripClusterizerAlgorithmFactory::create(consumesCollector(), conf);
 
   produces<output_t>("");
 }
 
-StripByStripTestDriver::~StripByStripTestDriver() {
-  //if(hltFactory) delete hltFactory;
-}
+StripByStripTestDriver::~StripByStripTestDriver() = default;
 
 void StripByStripTestDriver::produce(edm::Event& event, const edm::EventSetup& es) {
   auto output = std::make_unique<output_t>();
@@ -42,3 +62,6 @@ void StripByStripTestDriver::produce(edm::Event& event, const edm::EventSetup& e
   edm::LogInfo("Output") << output->dataSize() << " clusters from " << output->size() << " modules";
   event.put(std::move(output));
 }
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(StripByStripTestDriver);
